fix: removed-the-setTimeout-function
