+++ apps/api/src/app/country/country.service.ts
@@ -14,35 +14,4 @@ export class CountryService extends CrudService<Country> {
 	) {
 		super(countryRepository);
 	}
-	async getAsCsv() {
-		await fs.access('./export/csv', (error) => {
-			if (!error) {
-				return null;
-			} else {
-				fs.mkdir('./export/csv', { recursive: true }, (err) => {
-					if (err) throw err;
-				});
-			}
-		});
-
-		const createCsvWriter = csv.createObjectCsvWriter;
-		const dataIn = [];
-		const incommingData = (await this.findAll()).items;
-		const dataKeys = Object.keys(incommingData[0]);
-
-		for (const count of dataKeys) {
-			dataIn.push({ id: count, title: count });
-		}
-
-		const csvWriter = createCsvWriter({
-			path: './export/csv/country.csv',
-			header: dataIn
-		});
-
-		const data = incommingData;
-
-		csvWriter
-			.writeRecords(data)
-			.then(() => console.log('The CSV file was written successfully'));
-	}
 }
