+++ apps/api/src/app/export_import/export-all.module.ts
@@ -4,11 +4,12 @@ import { ExportAllController } from './export-all.controller';
 import { ExportAllService } from './export-all.service';
 import { CountryService, Country } from '../country';
 import { TypeOrmModule } from '@nestjs/typeorm';
+import { UserService, User } from '../user';
 
 @Module({
-	imports: [CqrsModule, TypeOrmModule.forFeature([Country])],
+	imports: [CqrsModule, TypeOrmModule.forFeature([Country, User])],
 	controllers: [ExportAllController],
-	providers: [ExportAllService, CountryService],
-	exports: [ExportAllService, CountryService]
+	providers: [ExportAllService, CountryService, UserService],
+	exports: [ExportAllService, CountryService, UserService]
 })
 export class ExportAllModule {}
